requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
spacy
networkx
alembic
pytest
